Remove duplicated endorsements and refactor
